r1 styles complete
